test/unit/modules/proposal/assets/document_alteration_asset.spec.ts
import { DocumentAlterationAsset } from '../../../../../src/app/modules/proposal/assets/document_alteration_asset';

describe('DocumentAlterationAsset', () => {
  let transactionAsset: DocumentAlterationAsset;

	beforeEach(() => {
		transactionAsset = new DocumentAlterationAsset();
	});

	describe('constructor', () => {
		it('should have valid id', () => {
			expect(transactionAsset.id).toEqual(2);
		});

		it('should have valid name', () => {
			expect(transactionAsset.name).toEqual('documentAlteration');
		});

		it('should have valid schema', () => {
			expect(transactionAsset.schema).toMatchSnapshot();
		});
	});

	describe('validate', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('apply', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
